Task flow analysis edits
Here is a first take. Trying to thread the needle by not referring to use cases (although, wondering if that use case ever come up as a card folks want? It is a bit engineery, and I know some designers are biased against the term, but as I see it, it is really the highest tier in the three part hierarchy of use case, user scenario, and task)...